LibWeb: Split NumericStyleValue out of StyleValue.{h,cpp}
