+++ Userland/Libraries/LibWeb/CSS/StyleValue.cpp
@@ -39,6 +39,7 @@
 #include <LibWeb/CSS/StyleValues/LengthStyleValue.h>
 #include <LibWeb/CSS/StyleValues/LinearGradientStyleValue.h>
 #include <LibWeb/CSS/StyleValues/ListStyleStyleValue.h>
+#include <LibWeb/CSS/StyleValues/NumericStyleValue.h>
 #include <LibWeb/CSS/StyleValues/RadialGradientStyleValue.h>
 #include <LibWeb/DOM/Document.h>
 #include <LibWeb/HTML/BrowsingContext.h>
@@ -1134,14 +1135,6 @@ ErrorOr<void> PositionValue::serialize(StringBuilder& builder) const
             return builder.try_append(TRY(length_percentage.to_string()));
         }));
     return {};
-}
-
-ErrorOr<String> NumericStyleValue::to_string() const
-{
-    return m_value.visit(
-        [](auto value) {
-            return String::formatted("{}", value);
-        });
 }
 
 ErrorOr<String> OverflowStyleValue::to_string() const
